// interfaces/SetUpCharterData.interface.ts
export interface SetUpCharterData {
    charterPerHour: number //1000
    chainteringServicePayPerHour: number //3
    minimumCruisingSpeed: number // 12
    vesselIMOnumber: number //9751779
    penaltyPerHour: number //20
    consuptionstandBy: number //5
    consuptionAtOperation: number //25
    consuptionUnderWay: number //20
}